requirements.txt
runpod>=1.6.0
requests>=2.32.0
boto3>=1.34.0
torch>=2.2.0
torchaudio>=2.2.0

# Wan I2V
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
sentencepiece>=0.2.0
Pillow>=10.0.0

# Real-ESRGAN
realesrgan>=0.3.0
basicsr>=1.4.2
opencv-python-headless>=4.8.0

# Wav2Lip + GFPGAN
gfpgan>=1.3.8

# Whisper
openai-whisper>=20231117

# AudioCraft (MusicGen + AudioGen)
audiocraft>=1.3.0
